LPD-75170 refactor: rename ai-hub-client-rest-client to ai-hub-cell-rest-client
