+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-client/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub Cell REST Client
+Bundle-SymbolicName: com.liferay.ai.hub.cell.rest.client
+Bundle-Version: 1.0.6
