+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-client/build.gradle
