+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-client/src/main/java/com/liferay/ai/hub/cell/rest/client/function/UnsafeSupplier.java
@@ -1,0 +1,20 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.cell.rest.client.function;
+
+import jakarta.annotation.Generated;
+
+/**
+ * @author Feliphe Marinho
+ * @generated
+ */
+@FunctionalInterface
+@Generated("")
+public interface UnsafeSupplier<T, E extends Throwable> {
+
+	public T get() throws E;
+
+}
