+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub Cell REST Implementation
+Bundle-SymbolicName: com.liferay.ai.hub.cell.rest.impl
+Bundle-Version: 1.0.6
