+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-impl/build.gradle
