+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-impl/rest-openapi.yaml
@@ -1,0 +1,41 @@
+components:
+    schemas:
+        Token:
+            properties:
+                accessToken:
+                    readOnly: true
+                    type: string
+                scope:
+                    readOnly: true
+                    type: string
+                serviceURL:
+                    readOnly: true
+                    type: string
+                userToken:
+                    readOnly: true
+                    type: string
+            type: object
+info:
+    description:
+        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
+        'com.liferay.ai.hub.cell.rest.client', and version '1.0.6'."
+    license:
+        name: Apache 2.0
+        url: http://www.apache.org/licenses/LICENSE-2.0.html
+    title: ""
+    version: v1.0
+openapi: 3.0.1
+paths:
+    "/tokens":
+        post:
+            operationId: postToken
+            responses:
+                200:
+                    content:
+                        application/json:
+                            schema:
+                                $ref: "#/components/schemas/Token"
+                        application/xml:
+                            schema:
+                                $ref: "#/components/schemas/Token"
+            tags: ["Token"]
