+++ modules/apps/ai-hub-cell/ai-hub-cell-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/token.properties
