--- modules/apps/ai-hub-cell/ai-hub-client-rest-client/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay AI Hub Client REST Client
-Bundle-SymbolicName: com.liferay.ai.hub.client.rest.client
-Bundle-Version: 1.0.6
