--- modules/apps/ai-hub-cell/ai-hub-client-rest-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay AI Hub Client REST Implementation
-Bundle-SymbolicName: com.liferay.ai.hub.client.rest.impl
-Bundle-Version: 1.0.6
