+++ modules/apps/ai-hub-cell/test.properties
@@ -1,4 +1,4 @@
-modified.files.includes[relevant][ai-hub-java-rule]=\
+modified.files.includes[relevant][ai-hub-cell-java-rule]=\
     **/*.java,\
     **/test/**,\
     **/testIntegration/**
