+++ test.properties
@@ -8050,7 +8050,7 @@
 
     testray.team.ai-hub.component.names=\
         AI Hub,\
-        AI Hub Client
+        AI Hub Cell
 
     testray.team.analytics-cloud.component.names=\
         A/B Test,\
